hashcode generation int overflow bug fix

+++ automata-and-logic-engineering-server/src/main/java/ale/server/ale/calculations/TruthTableBuilder.java
@@ -163,8 +163,8 @@ class TruthTableBuilder {
         for (int i = truthTable.size() - 1; i >= 0; i--) {
             binaryString += truthTable.get(i).get('=');
         }
-        int decimal = Integer.parseInt(binaryString, 2);
-        String hexString = Integer.toString(decimal, 16);
+        long decimal = Long.parseLong(binaryString, 2);
+        String hexString = Long.toString(decimal, 16);
         return hexString;
     }
 }
